microservices/user-service/requirements.txt
# Shared Library
-e ../shared_lib[redis,kafka]

# FastAPI
fastapi==0.116.1
uvicorn==0.35.0
gunicorn==23.0.0
python-multipart==0.0.20

# Database
aiomysql==0.2.0
sqlalchemy==2.0.43
redis==5.0.1
cryptography==44.0.0

# Kafka
aiokafka==0.11.0
python-snappy==0.7.3

# Authentication
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0

# Configuration
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
email-validator==2.1.0

# Utilities
aiofiles==25.1.0
httpx>=0.25.0
requests>=2.31.0

# Metrics
prometheus-fastapi-instrumentator==6.1.0
